fix(BJS-53035) Fix reviewed issues
