+++ src/main/java/faang/school/projectservice/client/JiraClient.java
@@ -1,5 +1,6 @@
 package faang.school.projectservice.client;
 
+import faang.school.projectservice.config.JiraConfig;
 import faang.school.projectservice.dto.jira.request.JiraIssueRequest;
 import faang.school.projectservice.dto.jira.response.JiraIssueResponse;
 import faang.school.projectservice.dto.jira.response.JiraSearchResponse;
@@ -7,22 +8,22 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-@FeignClient(name = "jiraClient", url = "${jira.url}")
+@FeignClient(name = "jiraClient", url = "${jira.url}", configuration = JiraConfig.class)
 public interface JiraClient {
 
-    @PostMapping("/rest/api/2/issue")
+    @PostMapping("/issue")
     ResponseEntity<JiraIssueResponse> createIssue(@RequestBody JiraIssueRequest request);
 
-    @PutMapping("/rest/api/2/issue/{issueIdOrKey}")
+    @PutMapping("/issue/{issueIdOrKey}")
     ResponseEntity<Void> updateIssue(
             @PathVariable("issueIdOrKey") String issueIdOrKey,
             @RequestBody JiraIssueRequest request
     );
 
-    @GetMapping("/rest/api/2/issue/{issueIdOrKey}")
+    @GetMapping("/issue/{issueIdOrKey}")
     ResponseEntity<JiraIssueResponse> getIssue(@PathVariable("issueIdOrKey") String issueId);
 
-    @GetMapping("/rest/api/2/search")
+    @GetMapping("/search")
     ResponseEntity<JiraSearchResponse> searchIssues(
             @RequestParam("jql") String jql,
             @RequestParam(value = "startAt", required = false, defaultValue = "0") Integer startAt,
