+++ src/main/java/faang/school/projectservice/controller/JiraController.java
@@ -9,7 +9,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/api/jira")
+@RequestMapping("${domain.base-path}/jira")
 @RequiredArgsConstructor
 public class JiraController {
