+++ src/main/java/faang/school/projectservice/dto/client/PaymentRequest.java
@@ -1,11 +1,11 @@
 package faang.school.projectservice.dto.client;
 
-import java.math.BigDecimal;
-
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
-public record PaymentRequest (
+import java.math.BigDecimal;
+
+public record PaymentRequest(
         Long paymentNumber,
 
         @Min(1)
