+++ src/main/java/faang/school/projectservice/dto/client/UserDto.java
@@ -1,8 +1,8 @@
 package faang.school.projectservice.dto.client;
 
-public record UserDto (
-    Long id,
-    String username,
-    String email
+public record UserDto(
+        Long id,
+        String username,
+        String email
 ) {
 }
