+++ src/main/java/faang/school/projectservice/dto/client/project/ProjectDto.java
@@ -1,7 +1,7 @@
 package faang.school.projectservice.dto.client.project;
 
 public record ProjectDto(
-    long id,
-    String title
+        long id,
+        String title
 ) {
 }
